Set A2A Password
Fixes issue #189. When calling the `PUT` method of the A2A `Credentials/Password` API, the code was taking the password value and blindly wrapping it in double quotes to try and make a JSON string object. But that would fail if the password value had a double quote character itself or other things.

So now, we properly serialize the password string value to a JSON string, with proper escaping.